Beginning of engine refactoring; break out interrupt_stream
