+++ drone/src/agent/engines/docker/mod.rs
@@ -1,4 +1,9 @@
-use crate::config::{DockerConfig, DockerConnection};
+mod util;
+use self::util::{AllowNotFound, ContainerEvent, ContainerEventType};
+use crate::{
+    agent::{engines::docker::util::{make_exposed_ports, MinuteExt}, engine::Engine},
+    config::{DockerConfig, DockerConnection},
+};
 use anyhow::{anyhow, Result};
 use bollard::{
     auth::DockerCredentials,
@@ -7,13 +12,14 @@ use bollard::{
         StatsOptions, StopContainerOptions,
     },
     image::CreateImageOptions,
-    models::{EventMessage, HostConfig, PortBinding, ResourcesUlimits},
+    models::{HostConfig, PortBinding, ResourcesUlimits},
     system::EventsOptions,
     Docker, API_DEFAULT_VERSION,
 };
-use plane_core::{messages::agent::ResourceLimits, timing::Timer};
+use plane_core::{messages::agent::ResourceLimits, timing::Timer, types::BackendId};
 use std::{collections::HashMap, net::IpAddr, time::Duration};
 use tokio_stream::{wrappers::IntervalStream, Stream, StreamExt};
+use std::pin::Pin;
 
 /// The port in the container which is exposed.
 const CONTAINER_PORT: u16 = 8080;
@@ -27,120 +33,6 @@ pub struct DockerInterface {
     docker: Docker,
     runtime: Option<String>,
     network: Option<String>,
-}
-
-/// Helper trait for swallowing Docker not found errors.
-trait AllowNotFound {
-    /// Swallow a result if it is a success result or a NotFound; propagate it otherwise.
-    fn allow_not_found(self) -> Result<(), bollard::errors::Error>;
-}
-
-impl<T> AllowNotFound for Result<T, bollard::errors::Error> {
-    fn allow_not_found(self) -> Result<(), bollard::errors::Error> {
-        match self {
-            Ok(_) => Ok(()),
-            Err(bollard::errors::Error::DockerResponseServerError {
-                status_code: 404, message,
-            }) => {
-                tracing::warn!(?message, "Received 404 error from docker, possibly expected.");
-                Ok(())
-            },
-            Err(e) => Err(e),
-        }
-    }
-}
-
-/// The list of possible container events.
-/// Comes from [Docker documentation](https://docs.docker.com/engine/reference/commandline/events/).
-#[derive(Debug, PartialEq, Eq)]
-pub enum ContainerEventType {
-    Attach,
-    Commit,
-    Copy,
-    Create,
-    Destroy,
-    Detach,
-    Die,
-    ExecCreate,
-    ExecDetach,
-    ExecDie,
-    ExecStart,
-    Export,
-    HealthStatus,
-    Kill,
-    Oom,
-    Pause,
-    Rename,
-    Resize,
-    Restart,
-    Start,
-    Stop,
-    Top,
-    Unpause,
-    Update,
-}
-
-#[allow(unused)]
-#[derive(Debug)]
-pub struct ContainerEvent {
-    pub event: ContainerEventType,
-    pub name: String,
-}
-
-impl ContainerEvent {
-    pub fn from_event_message(event: &EventMessage) -> Option<Self> {
-        let action = event.action.as_deref()?;
-        let actor = event.actor.as_ref()?;
-        let name: String = actor.attributes.as_ref()?.get("name")?.to_string();
-
-        let event = match action {
-            "attach" => ContainerEventType::Attach,
-            "commit" => ContainerEventType::Commit,
-            "copy" => ContainerEventType::Copy,
-            "create" => ContainerEventType::Create,
-            "destroy" => ContainerEventType::Destroy,
-            "detach" => ContainerEventType::Detach,
-            "die" => ContainerEventType::Die,
-            "exec_create" => ContainerEventType::ExecCreate,
-            "exec_detach" => ContainerEventType::ExecDetach,
-            "exec_die" => ContainerEventType::ExecDie,
-            "exec_start" => ContainerEventType::ExecStart,
-            "export" => ContainerEventType::Export,
-            "health_status" => ContainerEventType::HealthStatus,
-            "kill" => ContainerEventType::Kill,
-            "oom" => ContainerEventType::Oom,
-            "pause" => ContainerEventType::Pause,
-            "rename" => ContainerEventType::Rename,
-            "resize" => ContainerEventType::Resize,
-            "restart" => ContainerEventType::Restart,
-            "start" => ContainerEventType::Start,
-            "stop" => ContainerEventType::Stop,
-            "top" => ContainerEventType::Top,
-            "unpause" => ContainerEventType::Unpause,
-            "update" => ContainerEventType::Update,
-            _ => {
-                tracing::info!(?action, "Unhandled container action.");
-                return None;
-            }
-        };
-
-        Some(ContainerEvent { event, name })
-    }
-}
-
-trait MinuteExt {
-    fn as_minutes(&self) -> u128;
-}
-
-impl MinuteExt for std::time::Duration {
-    fn as_minutes(&self) -> u128 {
-        (self.as_secs() / 60).into()
-    }
-}
-
-fn make_exposed_ports(port: u16) -> Option<HashMap<String, HashMap<(), ()>>> {
-    let dummy: HashMap<(), ()> = vec![].into_iter().collect();
-    Some(vec![(format!("{}/tcp", port), dummy)].into_iter().collect())
 }
 
 impl DockerInterface {
@@ -163,23 +55,6 @@ impl DockerInterface {
             runtime: config.runtime.clone(),
             network: config.network.clone(),
         })
-    }
-
-    pub async fn container_events(&self) -> impl Stream<Item = ContainerEvent> {
-        let options: EventsOptions<&str> = EventsOptions {
-            since: None,
-            until: None,
-            filters: vec![("type", vec!["container"])].into_iter().collect(),
-        };
-        self.docker
-            .events(Some(options))
-            .filter_map(|event| match event {
-                Ok(event) => ContainerEvent::from_event_message(&event),
-                Err(error) => {
-                    tracing::error!(?error, "Error tracking container terminations.");
-                    None
-                }
-            })
     }
 
     pub fn get_logs(
@@ -412,5 +287,33 @@ impl DockerInterface {
         };
 
         Ok(())
+    }
+}
+
+impl Engine for DockerInterface {
+     fn interrupt_stream(&self) -> Pin<Box<dyn Stream<Item=plane_core::types::BackendId> + Send>> {
+        let options: EventsOptions<&str> = EventsOptions {
+            since: None,
+            until: None,
+            filters: vec![("type", vec!["container"])].into_iter().collect(),
+        };
+        let stream = self.docker
+            .events(Some(options))
+            .filter_map(|event| match event {
+                Ok(event) => {
+                    let event = ContainerEvent::from_event_message(&event)?;
+                    if event.event == ContainerEventType::Die {
+                        BackendId::from_resource_name(&event.name)
+                    } else {
+                        None
+                    }
+                },
+                Err(error) => {
+                    tracing::error!(?error, "Error tracking container terminations.");
+                    None
+                }
+            });
+
+        Box::pin(stream)
     }
 }
